+++ packages/classic-webview/src/hooks/useHardcoreAccess.tsx
@@ -1,7 +1,6 @@
 import { HardcoreAccessStatus } from '@hotandcold/classic-shared';
 import { createContext, useContext, useEffect, useState } from 'react';
 import { useDevvitListener } from './useDevvitListener';
-import { useModal } from './useModal';
 
 type HardcoreAccessContext = {
   access: HardcoreAccessStatus;
@@ -13,22 +12,12 @@ const hardcoreAccessContext = createContext<HardcoreAccessContext | null>(null);
 export const HardcoreAccessContextProvider = (props: { children: React.ReactNode }) => {
   const [access, setAccess] = useState<HardcoreAccessStatus>({ status: 'inactive' });
   const hardcoreAccessInitResponse = useDevvitListener('HARDCORE_ACCESS_INIT_RESPONSE');
-  const productPurchaseResponse = useDevvitListener('PURCHASE_PRODUCT_SUCCESS_RESPONSE');
-  const { closeModal } = useModal();
 
   useEffect(() => {
     if (hardcoreAccessInitResponse?.hardcoreAccessStatus != null) {
       setAccess(hardcoreAccessInitResponse.hardcoreAccessStatus);
     }
   }, [hardcoreAccessInitResponse, setAccess]);
-
-  // When a purchase is successful, update state and close the "Unlock Hardcore" modal
-  useEffect(() => {
-    if (productPurchaseResponse != null) {
-      setAccess(productPurchaseResponse.access);
-      closeModal();
-    }
-  }, [productPurchaseResponse, setAccess, closeModal]);
 
   return (
     <hardcoreAccessContext.Provider value={{ access, setAccess }}>
